mcp/requirements.txt
# Working requirements for Databricks Apps MCP Server
# Based on successful minimal deployment + MCP functionality

# Core FastAPI (updated for compatibility)
# fastapi>=0.116.1
fastapi
# uvicorn>=0.35.0
uvicorn
# pydantic>=2.11.7
pydantic
# databricks-sdk==0.65.0
databricks-sdk
# python-dotenv==1.0.1
python-dotenv
# pyyaml>=6.0.1
pyyaml

# MCP Framework (add back carefully)
fastmcp==2.3.2
# httpx==0.28.1

# DAB Generation Dependencies (essential only)
# astunparse>=1.6
# sqlparse>=0.4
# jinja2>=3.1
# jsonschema>=4.0
